moved perceptual models logic from theleafmap component into the mapview view
