+++ frontend/src/views/MapView.vue
@@ -59,9 +59,22 @@ import TheLeafletMap from '@/components/TheLeafletMap.vue'
 import { mdiChevronRight, mdiChevronLeft, mdiInformationOutline } from '@mdi/js'
 import { useMapStore } from '@/stores/map'
 import { useDisplay } from 'vuetify'
+import { storeToRefs } from 'pinia'
+import L from 'leaflet'
+
+const emit = defineEmits(['onFilter'])
+const mapStore = useMapStore()
+const {
+  mapLoaded,
+  userTouchedFilter,
+  currentFilteredData,
+  selectedSpatialZones,
+  selectedTemporalZones,
+  selectedProcesses,
+  searchTerm
+} = storeToRefs(mapStore)
 
 const { mdAndDown } = useDisplay()
-const mapStore = useMapStore()
 
 const showFilterDrawer = ref(true)
 const dataDrawerRef = ref(null)
@@ -72,8 +85,193 @@ watch(mdAndDown, (val) => {
   showDataDrawer.value = !val
 })
 
-onMounted(() => {
-  showFilterDrawer.value = !mdAndDown.value
+showFilterDrawer.value = !mdAndDown.value
+
+onMounted(async () => {
+  await mapStore.fetchPerceptualModelsGeojson()
+  const bounds = L.latLngBounds(mapStore.allAvailableCoordinates)
+  mapStore.leaflet.setMaxBounds(bounds)
+
+  const drawnItems = new L.FeatureGroup()
+  mapStore.drawnItems = drawnItems
+  mapStore.leaflet.addLayer(drawnItems)
+  drawnItems.setZIndex(1000)
+
+  let currentRectangle = null
+
+  L.Control.RectangleToggle = L.Control.extend({
+    options: { position: 'topleft' },
+    onAdd: function () {
+      const container = L.DomUtil.create(
+        'div',
+        'leaflet-bar leaflet-control leaflet-control-custom draw-toggle-btn'
+      )
+      updateDrawButton(container)
+      let drawer = null
+      L.DomEvent.on(container, 'click', () => {
+        if (currentRectangle || drawer) {
+          drawnItems.clearLayers()
+          currentRectangle = null
+          mapStore.filterFeatures(
+            (feature) => {
+              if (feature.geometry.type === 'Point') {
+                const [lng, lat] = feature.geometry.coordinates
+                return currentRectangle.getBounds().contains([lat, lng])
+              }
+              return false
+            },
+            'remove',
+            'rectangle'
+          )
+          userTouchedFilter.value = false
+          emit('onFilter', {
+            selectedSpatialZones,
+            selectedTemporalZones,
+            selectedProcesses,
+            searchTerm,
+            filteredFeatures: currentFilteredData.value
+          })
+          updateDrawButton(container)
+        } else {
+          const drawer = new L.Draw.Rectangle(mapStore.leaflet, {
+            shapeOptions: {
+              color: '#3388ff',
+              weight: 2,
+              opacity: 0.8,
+              fillOpacity: 0.3
+            },
+            showArea: false
+          })
+          drawer.enable()
+          currentRectangle = {}
+          updateDrawButton(container)
+
+          const drawHandler = (e) => {
+            mapStore.leaflet.off(L.Draw.Event.CREATED, drawHandler)
+            drawnItems.clearLayers()
+            currentRectangle = e.layer
+            currentRectangle.feature = {
+              type: 'Feature',
+              geometry: {
+                type: 'Polygon',
+                coordinates: [
+                  currentRectangle.getLatLngs()[0].map((latLng) => [latLng.lng, latLng.lat])
+                ]
+              },
+              properties: {}
+            }
+            drawnItems.addLayer(currentRectangle)
+            mapStore.leaflet.fitBounds(currentRectangle.getBounds())
+
+            mapStore.filterFeatures(
+              (feature) => {
+                if (feature.geometry.type === 'Point') {
+                  const [lng, lat] = feature.geometry.coordinates
+                  return currentRectangle.getBounds().contains([lat, lng])
+                }
+                return false
+              },
+              'add',
+              'rectangle'
+            )
+
+            userTouchedFilter.value = true
+            emit('onFilter', {
+              selectedSpatialZones,
+              selectedTemporalZones,
+              selectedProcesses,
+              searchTerm,
+              filteredFeatures: currentFilteredData.value
+            })
+
+            updateDrawButton(container)
+          }
+
+          mapStore.leaflet.on(L.Draw.Event.CREATED, drawHandler)
+          mapStore.leaflet.once('draw:drawstop', () => {
+            // If no rectangle was drawn, reset the toggle state
+            if (!drawnItems.getLayers().length) {
+              currentRectangle = null
+              updateDrawButton(container)
+              mapStore.leaflet.off(L.Draw.Event.CREATED, drawHandler)
+            }
+          })
+        }
+      })
+
+      return container
+    }
+  })
+  mapStore.leaflet.addControl(new L.Control.RectangleToggle())
+
+  async function mapClick() {
+    return
+  }
+  function updateDrawButton(container) {
+    if (currentRectangle) {
+      container.innerHTML = '<span class="material-icons">close</span>'
+      container.style.background = 'white'
+      container.title = 'Clear box'
+    } else {
+      container.style.backgroundImage = "url('/DrawIcon.ico')"
+      container.innerHTML = ''
+      container.title = 'Draw a box'
+    }
+    container.style.backgroundRepeat = 'no-repeat'
+    container.style.backgroundSize = '60% 60%'
+    container.style.backgroundPosition = 'center'
+    container.style.borderRadius = '4px'
+    container.style.width = '34px'
+    container.style.height = '34px'
+    container.style.cursor = 'pointer'
+    container.style.display = 'flex'
+    container.style.alignItems = 'center'
+    container.style.justifyContent = 'center'
+  }
+  L.drawLocal.draw.handlers.rectangle.tooltip.start = 'Click and drag to draw a box'
+  L.Control.ClearFilters = L.Control.extend({
+    options: { position: 'topleft' },
+    onAdd: function () {
+      const container = L.DomUtil.create(
+        'div',
+        'leaflet-bar leaflet-control leaflet-control-custom'
+      )
+      container.title = 'Reset Filters'
+
+      container.style.backgroundImage = "url('/ClearFilter.ico')"
+      container.style.backgroundRepeat = 'no-repeat'
+      container.style.backgroundSize = '60% 60%'
+      container.style.backgroundColor = 'white'
+      container.style.backgroundPosition = 'center'
+      container.style.borderRadius = '4px'
+      container.style.width = '34px'
+      container.style.height = '34px'
+      container.style.cursor = 'pointer'
+      container.style.display = 'flex'
+      container.style.alignItems = 'center'
+      container.style.justifyContent = 'center'
+
+      L.DomEvent.on(container, 'click', () => {
+        mapStore.clearAllFilters()
+        emit('onFilter', {
+          selectedSpatialZones,
+          selectedTemporalZones,
+          selectedProcesses,
+          searchTerm,
+          filteredFeatures: currentFilteredData.value
+        })
+      })
+
+      return container
+    }
+  })
+  mapStore.leaflet.addControl(new L.Control.ClearFilters())
+
+  mapStore.leaflet.on('click', function (e) {
+    mapClick(e)
+  })
+
+  mapLoaded.value = true
 })
 
 const onFilter = (data) => {
